CAPSTONE 2
ABOUT
This is a simple "Lord of the rings game" game program with one hero, a prize and three enemies 
that move around on the screen in different directions and the hero 
dies if he comes in contact with the enemies and wins if he comes in contact
with the prize. 

BUILT WITH
Python and Pygame

INSTALLED
VS Code, python IDE and pygame.

USAGE
User can move the hero in all four directions to avoid collision with the 
three enemies that are moving in random directions
while user moves toward the prize ("the ring"). Colliding with
an enemy leads to user dying and colliding with the prize leads to user winning.

KNOWN ISSUES
None.

CONTACT
MTK Liphoko
FitKayDev